+++ src/pages/list/ListPage.jsx
@@ -16,7 +16,7 @@ function ListPage() {
   const [sort, setSort] = useState("latest"); 
 
   const [isLiked, setIsLiked] = useState(false);
-  const [isLikeLoading, setIsLikeLoading] = useState(false); 
+  const [isLikeLoading, setIsLikeLoading] = useState(false); // 중복 클릭 방지용);
 
   const sortOptions = [
     { value: "latest", label: "최신순" },
@@ -25,6 +25,7 @@ function ListPage() {
   ];
 
   const sortedItems = useMemo(() => {
+    // 플레이리스트 아이템이 없으면 빈 배열 반환
   if (!playlistItem || !playlistItem.items) {
    return [];
   }
@@ -32,17 +33,20 @@ function ListPage() {
   const itemsToSort = [...playlistItem.items];
 
   if (sort === "latest") {
+      // API에서 받아온 createdAt(아이템 추가일)  기준으로 내림차순 정렬
       itemsToSort.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt));
     } else if (sort === "likes") {
+      // API에서 받아온 likeCount(아이템 좋아요 수)  기준으로 내림차순 정렬
       itemsToSort.sort((a, b) => (b.likeCount || 0) - (a.likeCount || 0));
     } else if (sort === "title") {
+      // 책 제목(b_title) [cite: 14] 기준으로 오름차순(가나다순) 정렬
       itemsToSort.sort((a, b) => {
         const titleA = a.book?.title || "";
         const titleB = b.book?.title || "";
         return titleA.localeCompare(titleB);
       });
     }
-
+    // 정렬된 배열 반환
   return itemsToSort;
  }, [playlistItem, sort]);
 
@@ -77,7 +81,7 @@ function ListPage() {
     if (isLikeLoading) return;
     setIsLikeLoading(true);
 
-    const originalIsLiked = isLiked;  
+    const originalIsLiked = isLiked;  // 롤백을 위해 현재 상태를 백업
     setIsLiked(!originalIsLiked);
 
     try {
@@ -85,13 +89,15 @@ function ListPage() {
     } catch (error) {
       console.error("좋아요 처리 실패:", error);
       alert("좋아요 처리에 실패했습니다. 다시 시도해주세요.");
-      setIsLiked(originalIsLiked); 
+      setIsLiked(originalIsLiked); // 상태를 원상 복구
     } finally {
+      // 성공/실패 여부와 관계없이 로딩 상태 해제
       setIsLikeLoading(false);
     }
   };
 
   const handleShare = () => {
+    // 웹 표준 공유 API 사용
     if (navigator.share) {
       navigator
         .share({
@@ -102,12 +108,14 @@ function ListPage() {
         .then(() => console.log("공유 성공"))
         .catch((error) => console.log("공유 실패", error));
     } else {
+      // 공유 API 미지원 시 클립보드 복사
       navigator.clipboard.writeText(window.location.href);
       alert("링크가 클립보드에 복사되었습니다.");
     }
   };
 
   const handleGoBack = () => {
+    // ListPage로 이동
     navigate(`/`);
   };
 
@@ -124,26 +132,29 @@ function ListPage() {
           className="absolute px-4 py-2 z-20 text-[20px] font-semibold text-white transition-opacity hover:opacity-75">
             ←
           </button>
+          {/* 썸네일 카드 */}
             <ThumbnailCard 
               image={playlistItem.image} 
               title={playlistItem.title} 
             />
           </div>
-          <div className="px-[15px] py-[5px] flex flex-col gap-2">  
+          <div className="px-[15px] py-[5px] flex flex-col gap-3">  
+          {/* 태그 표시 */}
           <p className="font-bold text-[15px] px-1 text-[#828282]">
             {Array.isArray(playlistItem.tags) 
               ? playlistItem.tags.map(tag => `#${tag.name}`).join(' ') 
               : ''}
           </p>
           
+          {/* QuoteBox */}
           <QuoteBox class
             text={playlistItem.description} />
 
           <div className="flex justify-between items-center px-1">
             <HeartShareButton
               isLiked={isLiked}
-              onHeartClick={handleLikeToggle} 
-              onShareClick={handleShare}
+              onHeartClick={handleLikeToggle} // 하트 로직 연결
+              onShareClick={handleShare} // 공유 로직 연결
             />
             <div className="flex items-center gap-4">
               <SortDropdown
@@ -152,13 +163,14 @@ function ListPage() {
                 initialValue={sort}
                 onChange={setSort}
               />
-              
+              {/* 책 개수 표시 */}
               <span className="text-sm font-bold text-[15px] text-[#828282]">
                 책: {playlistItem.items?.length || 0}개
               </span>
             </div>
           </div>
           
+          {/* 플레이리스트에 포함된 책 목록 */}
           {sortedItems.length > 0 ? (
             <div className="grid grid-cols-3 gap-4">
               {sortedItems.map(item => (
